refactor: change likeCount type to Long and optimize SponsorUtil client and thumbnail validation logic

+++ app/src/main/java/com/junkfood/seal/ui/page/tools/VideoInfoDownloadPage.kt
@@ -525,7 +525,7 @@ private fun ResultPreviewCard(
     thumbnailUrl: String?,
     durationString: String?,
     viewCount: Long?,
-    likeCount: Int?,
+    likeCount: Long?,
     tagCount: Int,
     hasDescription: Boolean,
     isSaved: Boolean,
